+++ packages/abjad-convert/src/scripts/ugaritic/letters.ts
@@ -1,4 +1,7 @@
-export const letters = [
+import { Script } from '../Script'
+import { Abjad } from '../../types'
+
+const letters = [
 	'\uD800\uDF80', /** #0  𐎀 U+10380 UGARITIC LETTER ALPA */
 	'\uD800\uDF81', /** #1  𐎁 U+10381 UGARITIC LETTER BETA */
 	'\uD800\uDF82', /** #2  𐎂  U+10382 UGARITIC LETTER GAMLA */
@@ -29,9 +32,9 @@ export const letters = [
 	'\uD800\uDF9B', /** #27 𐎛 U+1039B UGARITIC LETTER I */
 	'\uD800\uDF9C', /** #28 𐎜 U+1039C UGARITIC LETTER U */
 	'\uD800\uDF9D', /** #29 𐎝 U+1039D UGARITIC LETTER SSU */
-]
+] as const
 
-export const Ug = {
+const Ug = {
 	Alpa: letters[0], // 𐎀
 	Beta: letters[1], // 𐎁
 	Gamla: letters[2], // 𐎂
@@ -63,3 +66,10 @@ export const Ug = {
 	U: letters[28], // 𐎜
 	Ssu: letters[29], // 𐎝
 }
+
+export const ugaritic = new Script(
+	Abjad.Ugaritic,
+	false,
+	letters,
+	Ug,
+)